pyproject.toml
[tool.poetry]
name = "natsclient2"
version = "0.1.0"
description = ""
authors = ["hieu"]

[tool.poetry.dependencies]
python = "^3.7"
asyncio-nats-client = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
